requirements.txt
anthropic>=0.39.0
pyyaml>=6.0
colorama>=0.4.6
python-dotenv>=1.0.0
faker>=22.0.0  # For diverse agent name generation

# Database dependencies
psycopg2-binary>=2.9.9
qdrant-client>=1.7.0

# AI services
openai>=1.12.0  # For GPT-4o-mini metadata extraction and embeddings

# Phase 1: Dynamic Multi-Agent System
scikit-learn>=1.3.0  # For similarity detection and classification
numpy>=1.24.0  # For embeddings and numerical operations

# Optional: For alternative embedding models
# cohere>=4.0.0  # For Cohere embeddings
# sentence-transformers>=2.2.0  # For local embeddings

# Autonomous Search & Research (v0.6.0)
aiohttp>=3.9.0  # Async HTTP client for content fetching
beautifulsoup4>=4.12.0  # HTML parsing
readability-lxml>=0.8.1  # Content extraction with Readability algorithm
html2text>=2020.1.16  # HTML to Markdown conversion
pytz>=2024.1  # Timezone support (fallback for Python < 3.9)
